refactor: nettests is now an internal package
I'm moving everything into internal packages since this ain't a library.

While there, remove typo that was causing a build breakage.
