+++ internal/nettests/groups.go
