+++ internal/nettests/http_invalid_request_line.go
