+++ internal/nettests/nettests.go
