+++ internal/nettests/telegram.go
